fix problem with emailing of tracebacks for internal users of enclone visual
