+++ enclone_visual/src/enclone_client.rs
@@ -31,6 +31,8 @@
 use crate::proto::{analyzer_client::AnalyzerClient, ClonotypeRequest, EncloneRequest};
 use crate::*;
 use enclone_core::parse_bsv;
+use enclone_core::prepare_for_apocalypse::*;
+use enclone_core::REMOTE_HOST;
 use gui::launch_gui;
 use io_utils::*;
 use itertools::Itertools;
@@ -165,85 +167,105 @@ pub async fn enclone_client(t: &Instant) -> Result<(), Box<dyn std::error::Error
     let mut filehost = String::new();
     let mut filehost_used = false;
     let mut auto_update = false;
-    if config_name.len() > 0 {
-        let mut config_file_contents = String::new();
-        if config_name.len() > 0 {
-            for (key, value) in env::vars() {
-                if key == "ENCLONE_CONFIG" {
-                    CONFIG_FILE.lock().unwrap().push(value.to_string());
-                    let hf = value.to_string();
-                    let mut filename = hf.clone();
-                    if hf.contains(":") {
-                        filehost = hf.before(":").to_string();
-                        filename = hf.after(":").to_string();
-                    }
-
-                    // If filename exists on this server, use it, and ignore filehost.
-
-                    if path_exists(&filename) {
-                        config_file_contents = std::fs::read_to_string(&filename).unwrap();
-
-                    // Otherwise, fetch the file from the host.
-                    } else if filehost.len() > 0 {
-                        filehost_used = true;
-                        let t = Instant::now();
-                        let o = Command::new("ssh")
-                            .arg(&filehost)
-                            .arg("cat")
-                            .arg(&filename)
-                            .output()
-                            .expect("failed to execute ssh cat");
-                        println!("\nssh cat to {} took {:.1} seconds", filehost, elapsed(&t));
-                        if o.status.code() != Some(0) {
-                            let m = String::from_utf8(o.stderr).unwrap();
-                            println!("\ntest ssh failed with error message =\n{}", m);
-                            println!(
-                                "Attempt to ssh to {} as specified by environment variable \
-                                ENCLONE_CONFIG failed.",
-                                filehost,
-                            );
-                            println!("Here are two possible explanations:");
-                            println!("1. The host is wrong.");
-                            println!(
-                                "2. You first need to do something to enable crossing \
-                                      a firewall."
-                            );
-                            println!("   If so, ask one of your colleagues how to do this.\n");
-                            std::process::exit(1);
-                        }
-                        config_file_contents = strme(&o.stdout).to_string();
-                    }
-                }
+    let mut internal = false;
+    let mut config_file_contents = String::new();
+    for (key, value) in env::vars() {
+        if key == "ENCLONE_CONFIG" {
+            CONFIG_FILE.lock().unwrap().push(value.to_string());
+            let hf = value.to_string();
+            let mut filename = hf.clone();
+            if hf.contains(":") {
+                filehost = hf.before(":").to_string();
+                filename = hf.after(":").to_string();
             }
-        }
-
-        // Get configuration.
-
-        if config_name.len() > 0 {
-            let prefix = format!("vis.{}.", config_name);
-            for line in config_file_contents.lines() {
-                if line == "visual_auto_update=true" {
-                    auto_update = true;
-                }
-                if line.starts_with(&prefix) {
-                    let def = line.after(&prefix);
-                    if def.contains("=") {
-                        config.insert(def.before("=").to_string(), def.after("=").to_string());
-                        found = true;
-                    }
-                }
+            if filehost.len() > 0 {
+                REMOTE_HOST.lock().unwrap().push(filehost.clone());
             }
-        }
-        if !found {
-            eprintln!(
-                "\nYou specified the configuration name {}, but the content of that configuration \
-                   was not found.\n",
-                config_name,
-            );
-            std::process::exit(1);
+
+            // If filename exists on this server, use it, and ignore filehost.
+
+            if path_exists(&filename) {
+                config_file_contents = std::fs::read_to_string(&filename).unwrap();
+
+            // Otherwise, fetch the file from the host.
+            } else if filehost.len() > 0 {
+                filehost_used = true;
+                let t = Instant::now();
+                let o = Command::new("ssh")
+                    .arg(&filehost)
+                    .arg("cat")
+                    .arg(&filename)
+                    .output()
+                    .expect("failed to execute ssh cat");
+                println!("\nssh cat to {} took {:.1} seconds", filehost, elapsed(&t));
+                if o.status.code() != Some(0) {
+                    let m = String::from_utf8(o.stderr).unwrap();
+                    println!("\ntest ssh failed with error message =\n{}", m);
+                    println!(
+                        "Attempt to ssh to {} as specified by environment variable \
+                        ENCLONE_CONFIG failed.",
+                        filehost,
+                    );
+                    println!("Here are two possible explanations:");
+                    println!("1. The host is wrong.");
+                    println!(
+                        "2. You first need to do something to enable crossing \
+                              a firewall."
+                    );
+                    println!("   If so, ask one of your colleagues how to do this.\n");
+                    std::process::exit(1);
+                }
+                config_file_contents = strme(&o.stdout).to_string();
+            }
         }
     }
 
+    // Get configuration.
+
+    for line in config_file_contents.lines() {
+        if line == "visual_auto_update=true" {
+            auto_update = true;
+        }
+        if line == "internal=true" {
+            internal = true;
+        }
+        if config_name.len() > 0 {
+            let prefix = format!("vis.{}.", config_name);
+            if line.starts_with(&prefix) {
+                let def = line.after(&prefix);
+                if def.contains("=") {
+                    config.insert(def.before("=").to_string(), def.after("=").to_string());
+                    found = true;
+                }
+            }
+        }
+    }
+    if config_name.len() > 0 && !found {
+        eprintln!(
+            "\nYou specified the configuration name {}, but the content of that configuration \
+               was not found.\n",
+            config_name,
+        );
+        std::process::exit(1);
+    }
+
+    // Set up proper tracebacks.
+
+    let mut bug_reports = "enclone@10xgenomics.com".to_string();
+    for i in 1..args.len() {
+        if args[i] == "BUG_REPORTS" {
+            bug_reports = "".to_string();
+        } else if args[i].starts_with("BUG_REPORTS=") {
+            bug_reports = args[i].after("BUG_REPORTS=").to_string();
+        }
+    }
+    for (key, value) in env::vars() {
+        if key == "ENCLONE_BUG_REPORTS" {
+            bug_reports = value.to_string();
+        }
+    }
+    prepare_for_apocalypse(&args, internal, &bug_reports);
+        
     // Determine if the server is remote.
 
     let remote = config.contains_key("REMOTE_HOST")
